moved the API calls to a new .py file
