+++ apiCalls.py
@@ -1,0 +1,50 @@
+# Http integration managment
+# generic verion of a http request, formats to request well.
+def request():
+	pass
+
+# generic version of a http response, formats to respond well.
+def response():
+	pass
+
+
+
+# Note Searching
+def listNotes():
+	#by created date, modified date, title
+	pass
+
+def searchNotes():
+	#by content, title, tags, date
+	pass
+
+
+
+# Note Classification Suite
+def addTag():
+	# Makes a list of tags that should be added to the title, does not need to check if tags exist
+	pass
+
+def deletetag():
+	# Delete a tag from a note
+	pass
+
+def listTags():
+	# Lists tags of all notes
+	pass 
+
+
+
+# Note creation suite
+def createNote():
+	#addTag()
+	pass 
+
+def deleteNote():
+	pass
+
+def addContent():
+	# This should add to an existing note
+	pass 
+
+
